Added Extender examples
These examples can be used for segmenting a bus into mutiple media, for
example for range extension or spanning wired and wireless with the same
bus.
Also show is a pair of extenders working together to tunnel SWBB traffic
back and forth through an Ethernet LAN using the LocalUDP strategy.