Added single_initiate_direction mode to EthernetLink, Surrogate example
Surrogate and RemoteWorker scheme introduced, allowing a remote device
to parttake in a SWBB (or other strategy) PJON bus via Ethernet.
EthernetLink extended with a new mode to facilitate this efficiently, in
a way suited to firewalls and port forwarding.

Single_socket and single_initiate_direction modes are not compiled by
default, a preprocessor directive is needed to include the code.

Changed the default keep_connection value from true to false for the
EthernetLink and EthernetTCP strategy, to allow for many-to-many
communication out of the box.