Removed inactive code in PJONRouterExtended.
No support for segmented buses attached to routers only reachable
through the routing table. It is a requirement that if a bus is
segmented, all the segments must be attached to the same router, or
reachable through the default gateway. If this feature is needed
(probably not), it can be added later.