Robustified startup of LUDP-SWBB switches
Instead of the traditional Arduino method of simply giving up if the DHCP request fails, the request will be repeated until success. So the network cable need not be plugged in when starting, and it will be more robust to power failures where the DHCP server needs time to boot and the DHCP request would time out and the switch would fail.

Also added code to maintain the DHCP lease so that the IP address will not be handed out to another device after expiry, causing an address conflict.